+++ sdk/servicebus/service-bus/test/stress/stress_fixedNumberOfClientsSingleQueue.ts
@@ -68,7 +68,10 @@ async function sendReceiveMessages(): Promise<void> {
   }
 }
 
-async function sendReceiveMessagesPerClient(sender: Sender, receiver: InternalReceiver): Promise<void> {
+async function sendReceiveMessagesPerClient(
+  sender: Sender,
+  receiver: InternalReceiver
+): Promise<void> {
   while (!isJobDone) {
     const message: SendableMessageInfo = {
       messageId: msgId,