+++ sdk/storage/storage-blob/src/utils/Mutex.ts
@@ -53,8 +53,8 @@ export class Mutex {
     });
   }
 
-  private static keys: {[key: string]: MutexLockStatus} = {};
-  private static listeners: {[key: string]: Callback[]} = {};
+  private static keys: { [key: string]: MutexLockStatus } = {};
+  private static listeners: { [key: string]: Callback[] } = {};
 
   private static onUnlockEvent(key: string, handler: Callback) {
     if (this.listeners[key] === undefined) {